filter amenity detail

+++ src/java/control/DeleteAmenityDetailConatroller.java
@@ -2,7 +2,6 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package control;
 
 import dal.AmenityDAO;
@@ -13,41 +12,45 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 /**
  *
  * @author admin
  */
-@WebServlet(name="DeleteAmenityDetailConatroller", urlPatterns={"/deleteAmenityDetail"})
+@WebServlet(name = "DeleteAmenityDetailConatroller", urlPatterns = {"/deleteAmenityDetail"})
 public class DeleteAmenityDetailConatroller extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         try (PrintWriter out = response.getWriter()) {
             /* TODO output your page here. You may use following sample code. */
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet DeleteAmenityDetailConatroller</title>");  
+            out.println("<title>Servlet DeleteAmenityDetailConatroller</title>");
             out.println("</head>");
             out.println("<body>");
-            out.println("<h1>Servlet DeleteAmenityDetailConatroller at " + request.getContextPath () + "</h1>");
+            out.println("<h1>Servlet DeleteAmenityDetailConatroller at " + request.getContextPath() + "</h1>");
             out.println("</body>");
             out.println("</html>");
         }
-    } 
+    }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -55,17 +58,28 @@ public class DeleteAmenityDetailConatroller extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
-       String roomID = request.getParameter("roomid");
-        System.out.println("RoomId ============="+roomID);
-        AmenityDAO amenityDao = new AmenityDAO();
-        amenityDao.deleteAmenityDetailByRoomNumber(roomID);
-    
-    response.sendRedirect("amenity-detail");
-    } 
+            throws ServletException, IOException {
+        HttpSession session = request.getSession(false);
+        if (session == null) {
+            response.sendRedirect("login.jsp");
+        } else {
+            int role = Integer.parseInt(String.valueOf(session.getAttribute("role")));
+            if (session.getAttribute("role") != null && role != 2) {
+                request.setAttribute("error", "Please sign in with receptionist account !");
+                request.getRequestDispatcher("login.jsp").forward(request, response);
+            }
+            String roomID = request.getParameter("roomid");
+            System.out.println("RoomId =============" + roomID);
+            AmenityDAO amenityDao = new AmenityDAO();
+            amenityDao.deleteAmenityDetailByRoomNumber(roomID);
 
-    /** 
+            response.sendRedirect("amenity-detail");
+        }
+    }
+
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -73,12 +87,13 @@ public class DeleteAmenityDetailConatroller extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
